move /cypress/common to /cypress/support
