+++ tests/cypress/support/views.js
