+++ tests/cypress/tests/RoleBindingPolicy_governance.spec.js
@@ -3,7 +3,7 @@
 
 /// <reference types="cypress" />
 import { describeT } from '../support/tagging'
-import { test_genericPolicyGovernance, test_applyPolicyYAML } from '../common/tests'
+import { test_genericPolicyGovernance, test_applyPolicyYAML } from '../support/tests'
 
 describeT('@extended RHACM4K-1722 - GRC UI: [P1][Sev1][policy-grc] RoleBindingPolicy governance', () => {
   test_genericPolicyGovernance('RoleBinding_policy_governance/policy-config.yaml', 'RoleBinding_policy_governance/violations-inform.yaml', 'RoleBinding_policy_governance/violations-enforce.yaml')
